Corrected ExpandedNetwork.java
Under some very specific circumstances when searching for stable motifs
some cycles could get deleted when they shouldn't have. Namely, if a
cycle contained composite node C (with inputs A and B) and node A, and
another cycle contained nodes A, B, and C, and the first was fully
contained in the second one, then no new cycles would be added (since
the new cycle already existed) but the old cycle would be deleted.